requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
alpaca-trade-api==1.4.3
gunicorn==20.1.0
python-dotenv==0.19.0